Integrated code for remote control and line following
I have added functions for both line following and remote control modes. Some variable names have also been changed within the main code for each of the line following loop, and the remote control loop so that the two code bases could be merged here. The button to change the car's operation mode, which is initially remote control, is button a.